requirements.txt
alembic==1.7.7
anyio==3.6.1
asgiref==3.5.2
autopep8==1.6.0
click==8.1.3
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
Mako==1.2.0
MarkupSafe==2.1.1
pycodestyle==2.8.0
pydantic==1.9.1
pyhumps==3.7.1
PyJWT==2.4.0
python-dotenv==0.20.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
toml==0.10.2
typing-extensions==4.2.0
uvicorn==0.17.6
zipp==3.8.0
